Refactors authentication flow with Bloc pattern
Improves the authentication flow by integrating a Bloc pattern for state management.
This change addresses issues with inconsistent state updates and simplifies the authentication logic in the sign-in and sign-up pages.

The AuthBloc is now injected into the AuthButton widget, providing a centralized way to manage the authentication state and trigger UI updates. Additionally, the AuthBloc is registered as a Factory in dependency injection, guaranteeing that new instances of AuthBloc are created upon request.

Adds logout functionality in user account page with confirmation dialog